Narrows return type for sign and signAsync.
If I have followed the code correctly, both of these functions are guaranteed to return a Signature with `recovery` set.  By updating the return type to reflect this, using these functions becomes easier as you don't have to unsafely ignore or check the undefined possibility on the signature immediately after you sign.